requirements.txt
Flask==3.0.0
requests==2.31.0
google-cloud-storage==2.14.0
moviepy==1.0.3
faster-whisper==1.0.3
google-auth==2.26.1
gunicorn==21.2.0
imageio==2.33.1
imageio-ffmpeg==0.4.9
google-generativeai>=0.7.0